requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pandas==2.2.2
tabulate==0.9.0       # <- for DataFrame.to_markdown
langchain==0.3.27
langchain-ibm>=0.1.13
langgraph==0.5.4
xgboost==3.0.2
python-multipart==0.0.20
pydantic==2.10.3
jinja2==3.1.2
aiofiles==23.2.1
langsmith==0.4.8
python-dotenv==1.1.1
